TST: test_rocket.py reorganized.
